Verify types in assignments
(cherry picked from commit b6683b06304c6f8da17fb641e4ceb69537b45b9e)

+++ core/trino-main/src/test/java/io/trino/sql/planner/iterative/rule/TestImplementTableFunctionSource.java
@@ -1217,7 +1217,7 @@ public class TestImplementTableFunctionSource
                                     // coerce column c for co-partitioning
                                     p.project(
                                             Assignments.builder()
-                                                    .put(c, new Reference(BIGINT, "c"))
+                                                    .put(c, new Reference(TINYINT, "c"))
                                                     .put(d, new Reference(BIGINT, "d"))
                                                     .put(cCoerced, new Cast(new Reference(BIGINT, "c"), INTEGER))
                                                     .build(),
